The easy challenge is now complete.
I had misinterpreted slightly the definition of a game, so this
refactors the code to meet reqs. Now it has all the stats reported at
the end as well. The only thing missing is the opportunity to just keep
playing forever, but I can't really be bothered with more user
interaction code.